Run module now builds dictionary for upload
Renamed full_directions to get_directions - better to invoke that on an
object. It's a method - should be a verb.

Finding closest pool now takes an optional list of potential pools.
Defaults to hardwired, but allows adding other pools.

Added log messages

Next stage upload method and make it appear on Google Sheets